fix containerization

+++ backend/src/config/create-key-pair.ts
@@ -3,6 +3,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 
 import { PRIVATE_KEY_FILE_NAME, PUBLIC_KEY_FILE_NAME } from './config.ts';
+import logger from './logger.ts';
 const __dirname = import.meta.dirname;
 function genKeyPair() {
   const keyPair = crypto.generateKeyPairSync('rsa', {
@@ -16,12 +17,19 @@ function genKeyPair() {
       format: 'pem',
     },
   });
+  if (
+    fs.existsSync(path.join(__dirname, `${PRIVATE_KEY_FILE_NAME}.pem`)) &&
+    fs.existsSync(path.join(__dirname, `${PUBLIC_KEY_FILE_NAME}.pem`))
+  ) {
+    logger.info("Key pair already exists. Skipping generation.");
+    return;
+  }
   try {
     fs.writeFileSync(path.join(__dirname, `${PUBLIC_KEY_FILE_NAME}.pem`), keyPair.publicKey);
     fs.writeFileSync(path.join(__dirname, `${PRIVATE_KEY_FILE_NAME}.pem`), keyPair.privateKey);
   } catch (error: unknown) {
     throw new Error(
-      `Error occured while generating key pair: ${error instanceof Error ? error.message : ''}`,
+      `Error occurred while generating key pair: ${error instanceof Error ? error.message : ''}`,
     );
   }
 }
